Pull out input
